move thanking robot cog to extension
