+++ duckbot/cogs/robot/__init__.py
@@ -1,0 +1,5 @@
+from .thanking_robot import ThankingRobot
+
+
+def setup(bot):
+    bot.add_cog(ThankingRobot(bot))
