+++ duckbot/cogs/robot/thanking_robot.py
